# src/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "Auth Server for Contact Tracing",
    "version": "1.0.0",
    "title": "Auth Server"
  },
  "tags": [
    {
      "name": "Authentication Service",
      "description": "All the necessary endpoints to ensure a private and secure authentication experience"
    }
  ],
  "paths": {
    "/users/signUp": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "create a new email and password user",
        "operationId": "usersSignUp",
        "requestBody": {
          "description": "Created user object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/usersSignUpRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/signUpResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing fields / Invalid visibility"
          },
          "409": {
            "description": "User already registered"
          }
        }
      }
    },
    "/owners/signUp": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "create a new email and password owner",
        "operationId": "ownersSignUp",
        "requestBody": {
          "description": "Created owner object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/ownersSignUpRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/signUpResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing fields / Invalid visibility"
          },
          "409": {
            "description": "User already registered"
          }
        }
      }
    },
    "/users/logIn": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Log in with email and password for role user",
        "operationId": "usersLogIn",
        "requestBody": {
          "description": "Login user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/logInRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/logInResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Email not found / Invalid password / User disabled"
          },
          "404": {
            "description": "El usuario logueado no tiene el rol user"
          }
        }
      }
    },
    "/owners/logIn": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Log in with email and password for role owner",
        "operationId": "ownersLogIn",
        "requestBody": {
          "description": "Login owner",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/logInRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/logInResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Email not found / Invalid password / User disabled"
          },
          "404": {
            "description": "El usuario logueado no tiene el rol owner"
          }
        }
      }
    },
    "/admin/logIn": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Log in with email and password for role admin",
        "operationId": "adminsLogIn",
        "requestBody": {
          "description": "Login user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/logInRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User logged in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/logInResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Email not found / Invalid password / User disabled"
          },
          "404": {
            "description": "El usuario logueado no tiene el rol admin"
          }
        }
      }
    },
    "/users/validateAccessToken": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Validate user ID token for role user",
        "operationId": "usersValidateAccessToken",
        "requestBody": {
          "description": "Validate Access token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/validateAccessTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/validateAccessTokenResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad formatted or invalid token"
          },
          "404": {
            "description": "El usuario logueado no tiene el rol user"
          }
        }
      }
    },
    "/owners/validateAccessToken": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Validate user ID token for role owner",
        "operationId": "ownersValidateAccessToken",
        "requestBody": {
          "description": "Validate Access token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/validateAccessTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/validateAccessTokenResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad formatted or invalid token"
          },
          "404": {
            "description": "El usuario logueado no tiene el rol owner"
          }
        }
      }
    },
    "/admins/validateAccessToken": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Validate user ID token for role admin",
        "operationId": "adminsValidateAccessToken",
        "requestBody": {
          "description": "Validate Access token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/validateAccessTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token is valid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/validateAccessTokenResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad formatted or invalid token"
          },
          "404": {
            "description": "El usuario logueado no tiene el rol admin"
          }
        }
      }
    },
    "/refreshToken": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Obtain a new ID Token providing the refresh Token",
        "operationId": "refreshToken",
        "requestBody": {
          "description": "Refresh Token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/refreshTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ID Token generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/refreshTokenResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Refresh Token expired, missing or invalid / Invalid password / User disabled or not found"
          }
        }
      }
    },
    "/generateGenuxToken": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Obtain a new Genux Token providing the ID Token",
        "operationId": "generateGenuxToken",
        "requestBody": {
          "description": "Generate Genux Token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/generateGenuxTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Genux Token generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/generateGenuxTokenResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "ID Token expired, missing or invalid"
          }
        }
      }
    },
    "/useGenuxToken": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "Use a Genux Token and disable it",
        "operationId": "useGenuxToken",
        "requestBody": {
          "description": "Use Genux Token",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/useGenuxTokenRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Genux Token used successfully"
          },
          "400": {
            "description": "Genux Token already used, invalid or does not exist"
          }
        }
      }
    },
    "/sendPasswordResetEmail": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "send an email to reset the password",
        "operationId": "sendPasswordResetEmail",
        "requestBody": {
          "description": "Send Password Reset Email",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/sendPasswordResetEmailRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/sendPasswordResetEmailResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing fields / Invalid visibility"
          }
        }
      }
    },
    "/confirmPasswordReset": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "confirm the password reset with the oob code received by email",
        "operationId": "confirmPasswordReset",
        "requestBody": {
          "description": "Confirm Password Reset",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/confirmPasswordResetRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password resetted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/confirmPasswordResetResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing fields / Invalid visibility"
          }
        }
      }
    },
    "/changePassword": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "change password of the current account while logged in",
        "operationId": "changePassword",
        "requestBody": {
          "description": "Change Password",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/changePasswordRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password changed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/changePasswordResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing fields / Invalid visibility"
          }
        }
      }
    },
    "/sendEmailVerification": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "send email to verify the account",
        "operationId": "sendEmailVerification",
        "requestBody": {
          "description": "Send Email Verification",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/sendEmailVerificationRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Account verified correctly",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/sendEmailVerificationResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing fields / Invalid visibility"
          }
        }
      }
    },
    "/getUserData": {
      "post": {
        "tags": [
          "Authentication Service"
        ],
        "summary": "get user data of the currently logged in use",
        "operationId": "getUserData",
        "requestBody": {
          "description": "Get User Data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "allOf": [
                  {
                    "$ref": "#/components/schemas/getUserDataRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User information retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/getUserDataResponse"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Missing fields / Invalid visibility"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "usersSignUpRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "securePassword"
          },
          "DNI": {
            "type": "string",
            "example": "40404040"
          }
        }
      },
      "ownersSignUpRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "securePassword"
          }
        }
      },
      "signUpResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "qoiu21noi75081jnln"
          },
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          },
          "refreshToken": {
            "type": "string",
            "example": "lkj12341injh98nv98"
          },
          "expiresIn": {
            "type": "string",
            "example": 3600
          },
          "userId": {
            "type": "string",
            "example": "tRcfmLH7n6082nlDF"
          }
        }
      },
      "logInRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "securePassword"
          }
        }
      },
      "logInResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "qoiu21noi75081jnln"
          },
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          },
          "refreshToken": {
            "type": "string",
            "example": "lkj12341injh98nv98"
          },
          "expiresIn": {
            "type": "string",
            "example": 3600
          },
          "userId": {
            "type": "string",
            "example": "tRcfmLH7n6082nlDF"
          }
        }
      },
      "validateAccessTokenRequest": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "asdf1234ghty"
          }
        }
      },
      "validateAccessTokenResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "tRcfmLH7n6082nlDF"
          }
        }
      },
      "refreshTokenRequest": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "lkj12341injh98nv98"
          }
        }
      },
      "refreshTokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "qoiu21noi75081jnln"
          },
          "expiresIn": {
            "type": "string",
            "example": 3600
          },
          "userId": {
            "type": "string",
            "example": "tRcfmLH7n6082nlDF"
          }
        }
      },
      "generateGenuxTokenRequest": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "qoiu21noi75081jnln"
          }
        }
      },
      "generateGenuxTokenResponse": {
        "type": "object",
        "properties": {
          "genuxToken": {
            "type": "string",
            "example": "pomnopiu1poi5poiu1212345"
          },
          "expiresIn": {
            "type": "string",
            "example": 3600
          }
        }
      },
      "useGenuxTokenRequest": {
        "type": "object",
        "properties": {
          "genuxToken": {
            "type": "string",
            "example": "pomnopiu1poi5poiu1212345"
          }
        }
      },
      "sendPasswordResetEmailRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          }
        }
      },
      "sendPasswordResetEmailResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          }
        }
      },
      "confirmPasswordResetRequest": {
        "type": "object",
        "properties": {
          "oobCode": {
            "type": "string",
            "example": "lkjovin1o234n5"
          },
          "newPassword": {
            "type": "string",
            "example": "securePassword"
          }
        }
      },
      "confirmPasswordResetResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          },
          "requestType": {
            "type": "string",
            "example": "PASSWORD_RESET"
          }
        }
      },
      "changePasswordRequest": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "qoiu21noi75081jnln"
          },
          "password": {
            "type": "string",
            "example": "securePassword"
          }
        }
      },
      "changePasswordResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "qoiu21noi75081jnln"
          },
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          },
          "refreshToken": {
            "type": "string",
            "example": "lkj12341injh98nv98"
          },
          "expiresIn": {
            "type": "string",
            "example": 3600
          },
          "userId": {
            "type": "string",
            "example": "tRcfmLH7n6082nlDF"
          }
        }
      },
      "sendEmailVerificationRequest": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "qoiu21noi75081jnln"
          }
        }
      },
      "sendEmailVerificationResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          }
        }
      },
      "getUserDataRequest": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "qoiu21noi75081jnln"
          }
        }
      },
      "getUserDataResponse": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "tomi@gmail.com"
          },
          "userId": {
            "type": "string",
            "example": "tRcfmLH7n6082nlDF"
          },
          "emailVerified": {
            "type": "boolean",
            "example": true
          },
          "displayName": {
            "type": "string",
            "example": "John Doe"
          },
          "photoUrl": {
            "type": "string",
            "example": "https://lh5.googleusercontent.com/.../photo.jpg"
          },
          "lastLoginAt": {
            "type": "string",
            "example": "1484628946000"
          },
          "createdAt": {
            "type": "string",
            "example": "1484124142000"
          },
          "DNI": {
            "type": "string",
            "example": "40404040"
          },
          "role": {
            "type": "string",
            "example": "regular"
          }
        }
      }
    }
  }
}